# backend/docs/swagger.yaml
basePath: /
definitions:
...
  domain.ObjectUploadResp:
    properties:
      filename:
        type: string
      key:
        type: string
    type: object
...
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  v1.FileUploadResp:
    properties:
      key:
        type: string
    type: object
...
info:
  contact: {}
  description: panda-wiki API documentation
  title: panda-wiki API
  version: "1.0"
paths:
...
  /api/v1/file/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload File
      parameters:
      - description: File
        in: formData
        name: file
        required: true
        type: file
      - description: Knowledge Base ID
        in: formData
        name: kb_id
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ObjectUploadResp'
      summary: Upload File
      tags:
      - file
...
  /share/v1/app/wechat/service/answer:
    get:
      consumes:
      - application/json
      description: GetWechatAnswer
      parameters:
      - description: conversation id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: GetWechatAnswer
      tags:
      - Wechat
...
  /share/v1/common/file/upload:
    post:
      consumes:
      - multipart/form-data
      description: 前台用户上传文件,目前只支持图片文件上传
      operationId: share-FileUpload
      parameters:
      - description: kb id
        in: header
        name: X-KB-ID
        required: true
        type: string
      - description: File
        in: formData
        name: file
        required: true
        type: file
      - description: captcha_token
        in: formData
        name: captcha_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.FileUploadResp'
              type: object
      summary: 文件上传
      tags:
      - ShareFile
...
securityDefinitions:
  bearerAuth:
    description: Type "Bearer" + a space + your token to authorize
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"